Вынес тесты из src/
