+++ test/api/HttpTaskManagerOfSubTasksTest.java
@@ -1,4 +1,4 @@
-package tests.Api;
+package api;
 
 import httpserver.HttpTaskServer;
 import com.google.gson.Gson;
@@ -12,7 +12,6 @@ import taskmanagers.TaskManager;
 import taskmanagers.TaskManagerFactory;
 import tasks.Epic;
 import tasks.SubTask;
-import tools.json.EpicTypeAdapter;
 import tools.json.SubTaskTypeAdapter;
 
 import java.io.IOException;
@@ -25,10 +24,9 @@ import java.time.Duration;
 import java.time.LocalDateTime;
 import java.util.List;
 
-public class HttpTaskManagerOfEpicsTest {
+public class HttpTaskManagerOfSubTasksTest {
     private final Gson gson = new GsonBuilder()
             .registerTypeAdapter(SubTask.class, new SubTaskTypeAdapter())
-            .registerTypeAdapter(Epic.class, new EpicTypeAdapter())
             .create();
 
     private TaskManager taskManager;
@@ -48,58 +46,75 @@ public class HttpTaskManagerOfEpicsTest {
 
     @Test
     void getList() throws IOException, InterruptedException, URISyntaxException {
-        taskManager.createEpic(new Epic("",""));
-        taskManager.createEpic(new Epic("",""));
-        taskManager.createEpic(new Epic("",""));
+        var epic = new Epic("","");
+        taskManager.createEpic(epic);
+
+        var epicId = epic.getTaskId();
+
+        taskManager.createSubTask(new SubTask("", "", epicId));
+        taskManager.createSubTask(new SubTask("", "", epicId));
 
         var response = getResponse(HttpRequest.newBuilder()
                 .GET()
-                .uri(new URI("http://127.0.0.1:8080/epics")));
+                .uri(new URI("http://127.0.0.1:8080/subtasks")));
 
         Assertions.assertEquals(200, response.statusCode(), response.body());
 
         var body = response.body();
 
-        List<Epic> tasks = gson.fromJson(body, new TypeToken<List<Epic>>() {}.getType());
+        List<SubTask> tasks = gson.fromJson(body, new TypeToken<List<SubTask>>() {}.getType());
 
-        Assertions.assertEquals(3, tasks.size());
+        Assertions.assertEquals(2, tasks.size());
     }
 
     @Test
     void noFoundTask() throws IOException, InterruptedException, URISyntaxException {
-        taskManager.createEpic(new Epic("",""));
-        taskManager.createEpic(new Epic("",""));
-        taskManager.createEpic(new Epic("",""));
+        var epic = new Epic("","");
+        taskManager.createEpic(epic);
+
+        var epicId = epic.getTaskId();
+
+        taskManager.createSubTask(new SubTask("", "", epicId));
+        taskManager.createSubTask(new SubTask("", "", epicId));
 
         var response = getResponse(HttpRequest.newBuilder()
                 .GET()
-                .uri(new URI("http://127.0.0.1:8080/epics/10")));
+                .uri(new URI("http://127.0.0.1:8080/subtasks/10")));
 
         Assertions.assertEquals(404, response.statusCode(), response.body());
     }
 
     @Test
     void getTask() throws IOException, InterruptedException, URISyntaxException {
-        taskManager.createEpic(new Epic("",""));
-        taskManager.createEpic(new Epic("",""));
         var epic = new Epic("","");
         taskManager.createEpic(epic);
-        taskManager.createEpic(new Epic("",""));
 
+        var epicId = epic.getTaskId();
 
-        taskManager.createSubTask(new SubTask("", "", epic.getTaskId()));
-        taskManager.createSubTask(new SubTask("", "", epic.getTaskId()));
+        taskManager.createSubTask(new SubTask("", "", epicId));
+        taskManager.createSubTask(new SubTask("", "", epicId));
+
+        epic = new Epic("","");
+        taskManager.createEpic(epic);
+
+        epicId = epic.getTaskId();
+
+        var subtask = new SubTask("subTask", "desc", epicId);
+
+        taskManager.createSubTask(new SubTask("", "", epicId));
+        taskManager.createSubTask(subtask);
+        taskManager.createSubTask(new SubTask("", "", epicId));
 
 
         var response = getResponse(HttpRequest.newBuilder()
                 .GET()
-                .uri(new URI("http://127.0.0.1:8080/epics/" + epic.getTaskId())));
+                .uri(new URI("http://127.0.0.1:8080/subtasks/" + subtask.getTaskId())));
 
         Assertions.assertEquals(200, response.statusCode(), response.body());
 
-        var actualTask = gson.fromJson(response.body(), Epic.class);
+        var actualTask = gson.fromJson(response.body(), SubTask.class);
 
-        Assertions.assertEquals(epic, actualTask);
+        Assertions.assertEquals(subtask, actualTask);
     }
 
     @Test
@@ -107,20 +122,46 @@ public class HttpTaskManagerOfEpicsTest {
         var epic = new Epic("","");
         taskManager.createEpic(epic);
 
-        epic = new Epic("epicNew","");
+        var subtask = new SubTask("subTask", "desc", epic.getTaskId());
+        subtask.setStartTime(LocalDateTime.of(2024,12,12,5,5, 34));
+        subtask.setDuration(Duration.ofHours(1));
+        subtask.doProgress();
 
         var response = getResponse(HttpRequest.newBuilder()
-                .POST(HttpRequest.BodyPublishers.ofString(gson.toJson(epic)))
-                .uri(new URI("http://127.0.0.1:8080/epics")));
+                .POST(HttpRequest.BodyPublishers.ofString(gson.toJson(subtask)))
+                .uri(new URI("http://127.0.0.1:8080/subtasks")));
 
         Assertions.assertEquals(201, response.statusCode(), response.body());
 
-        Assertions.assertEquals(2, taskManager.getEpics().size());
+        Assertions.assertEquals(1, taskManager.getSubTasks().size());
 
-        var newTask = taskManager.getEpics().get(1);
-        epic.setTaskId(newTask.getTaskId());
+        var newTask = taskManager.getSubTasks().getFirst();
+        subtask.setTaskId(newTask.getTaskId());
 
-        Assertions.assertEquals(epic, newTask);
+        Assertions.assertEquals(subtask, newTask);
+    }
+
+    @Test
+    void createCrossedTask() throws IOException, InterruptedException, URISyntaxException {
+        var epic = new Epic("","");
+        taskManager.createEpic(epic);
+
+        var subTask = new SubTask("task1","desc", epic.getTaskId());
+        subTask.setStartTime(LocalDateTime.of(2024,12,12,5,5));
+        subTask.setDuration(Duration.ofHours(1));
+        taskManager.createSubTask(subTask);
+
+        subTask = new SubTask("task2","desc", epic.getTaskId());
+        subTask.setStartTime(LocalDateTime.of(2024,12,12,5,55));
+        subTask.setDuration(Duration.ofHours(1));
+
+        var response = getResponse(HttpRequest.newBuilder()
+                .POST(HttpRequest.BodyPublishers.ofString(gson.toJson(subTask)))
+                .uri(new URI("http://127.0.0.1:8080/subtasks")));
+
+        Assertions.assertEquals(406, response.statusCode(), response.body());
+
+        Assertions.assertEquals(1, taskManager.getSubTasks().size());
     }
 
     @Test
@@ -130,15 +171,29 @@ public class HttpTaskManagerOfEpicsTest {
         if (res.isPresent())
             throw res.get();
 
-        epic.setName("updated");
+        var subTask = new SubTask("task1","desc", epic.getTaskId());
+        subTask.setStartTime(LocalDateTime.of(2024,12,12,5,5));
+        subTask.setDuration(Duration.ofHours(1));
+        res = taskManager.createSubTask(subTask);
+        if (res.isPresent())
+            throw res.get();
+
+        subTask = new SubTask("task2","desc", epic.getTaskId());
+        subTask.setStartTime(LocalDateTime.of(2024,12,12,6,55));
+        subTask.setDuration(Duration.ofHours(1));
+        res = taskManager.createSubTask(subTask);
+        if (res.isPresent())
+            throw res.get();
+
+        subTask.setName("updated");
 
         var response = getResponse(HttpRequest.newBuilder()
-                .POST(HttpRequest.BodyPublishers.ofString(gson.toJson(epic)))
-                .uri(new URI("http://127.0.0.1:8080/epics/" + epic.getTaskId())));
+                .POST(HttpRequest.BodyPublishers.ofString(gson.toJson(subTask)))
+                .uri(new URI("http://127.0.0.1:8080/subtasks/" + subTask.getTaskId())));
 
         Assertions.assertEquals(201, response.statusCode(), response.body());
 
-        var actualTask = taskManager.getEpic(epic.getTaskId());
+        var actualTask = taskManager.getSubTask(subTask.getTaskId());
         if (actualTask.isEmpty())
             Assertions.fail("не нашел task");
 
@@ -146,8 +201,46 @@ public class HttpTaskManagerOfEpicsTest {
     }
 
     @Test
+    void updateCrossTask() throws Exception {
+        var epic = new Epic("","");
+        var res = taskManager.createEpic(epic);
+        if (res.isPresent())
+            throw res.get();
+
+        var subTask = new SubTask("task1","desc", epic.getTaskId());
+        subTask.setStartTime(LocalDateTime.of(2024,12,12,5,5));
+        subTask.setDuration(Duration.ofHours(1));
+        res = taskManager.createSubTask(subTask);
+        if (res.isPresent())
+            throw res.get();
+
+        subTask = new SubTask("task2","desc", epic.getTaskId());
+        subTask.setStartTime(LocalDateTime.of(2024,12,12,6,55));
+        subTask.setDuration(Duration.ofHours(1));
+        res = taskManager.createSubTask(subTask);
+        if (res.isPresent())
+            throw res.get();
+
+        var response = getResponse(HttpRequest.newBuilder()
+                .GET()
+                .uri(new URI("http://127.0.0.1:8080/subtasks/" + subTask.getTaskId())));
+
+        var updatedTask = gson.fromJson(response.body(), SubTask.class);
+        updatedTask.setStartTime(LocalDateTime.of(2024,12,12,5,55));
+
+        response = getResponse(HttpRequest.newBuilder()
+                .POST(HttpRequest.BodyPublishers.ofString(gson.toJson(updatedTask)))
+                .uri(new URI("http://127.0.0.1:8080/subtasks/" + updatedTask.getTaskId())));
+
+        Assertions.assertEquals(406, response.statusCode(), response.body());
+
+        Assertions.assertNotEquals(subTask, updatedTask);
+    }
+
+
+    @Test
     void getPrioritizedList() throws Exception {
-        var epic = new Epic("1","");
+        var epic = new Epic("","");
         var res = taskManager.createEpic(epic);
         if (res.isPresent())
             throw res.get();
@@ -159,10 +252,6 @@ public class HttpTaskManagerOfEpicsTest {
         if (res.isPresent())
             throw res.get();
 
-        epic = new Epic("2","");
-        res = taskManager.createEpic(epic);
-        if (res.isPresent())
-            throw res.get();
         subTask = new SubTask("task2","desc", epic.getTaskId());
         subTask.setStartTime(LocalDateTime.of(2024,12,12,6,55));
         subTask.setDuration(Duration.ofMinutes(1));
@@ -170,10 +259,11 @@ public class HttpTaskManagerOfEpicsTest {
         if (res.isPresent())
             throw res.get();
 
-        epic = new Epic("3","");
+        epic = new Epic("","");
         res = taskManager.createEpic(epic);
         if (res.isPresent())
             throw res.get();
+
         subTask = new SubTask("task1","desc", epic.getTaskId());
         subTask.setStartTime(LocalDateTime.of(2024,12,12,3,5));
         subTask.setDuration(Duration.ofMinutes(1));
@@ -181,10 +271,6 @@ public class HttpTaskManagerOfEpicsTest {
         if (res.isPresent())
             throw res.get();
 
-        epic = new Epic("4","");
-        res = taskManager.createEpic(epic);
-        if (res.isPresent())
-            throw res.get();
         subTask = new SubTask("task2","desc", epic.getTaskId());
         subTask.setStartTime(LocalDateTime.of(2024,12,12,2,55));
         subTask.setDuration(Duration.ofMinutes(1));
@@ -194,23 +280,24 @@ public class HttpTaskManagerOfEpicsTest {
 
         var response = getResponse(HttpRequest.newBuilder()
                 .GET()
-                .uri(new URI("http://127.0.0.1:8080/epics/prioritized")));
+                .uri(new URI("http://127.0.0.1:8080/subtasks/prioritized")));
 
         Assertions.assertEquals(200, response.statusCode());
 
-        List<Epic> prioritizedList = gson.fromJson(response.body(), new TypeToken<List<Epic>>() {}.getType());
-        var names = new String[] {"4","3","1","2"};
+        List<SubTask> prioritizedList = gson.fromJson(response.body(), new TypeToken<List<SubTask>>() {}.getType());
+        var hours = new Integer[] {2,3,5,6};
         var index = 0;
 
         for (var item : prioritizedList) {
-            Assertions.assertEquals(names[index], item.getName());
+            var expectedVal = hours[index];
+            Assertions.assertTrue(item.getStartTime().filter(i -> i.getHour() == expectedVal).isPresent());
             index++;
         }
     }
 
     @Test
     void getPrioritizedDescList() throws Exception {
-        var epic = new Epic("1","");
+        var epic = new Epic("","");
         var res = taskManager.createEpic(epic);
         if (res.isPresent())
             throw res.get();
@@ -222,10 +309,6 @@ public class HttpTaskManagerOfEpicsTest {
         if (res.isPresent())
             throw res.get();
 
-        epic = new Epic("2","");
-        res = taskManager.createEpic(epic);
-        if (res.isPresent())
-            throw res.get();
         subTask = new SubTask("task2","desc", epic.getTaskId());
         subTask.setStartTime(LocalDateTime.of(2024,12,12,6,55));
         subTask.setDuration(Duration.ofMinutes(1));
@@ -233,10 +316,11 @@ public class HttpTaskManagerOfEpicsTest {
         if (res.isPresent())
             throw res.get();
 
-        epic = new Epic("3","");
+        epic = new Epic("","");
         res = taskManager.createEpic(epic);
         if (res.isPresent())
             throw res.get();
+
         subTask = new SubTask("task1","desc", epic.getTaskId());
         subTask.setStartTime(LocalDateTime.of(2024,12,12,3,5));
         subTask.setDuration(Duration.ofMinutes(1));
@@ -244,10 +328,6 @@ public class HttpTaskManagerOfEpicsTest {
         if (res.isPresent())
             throw res.get();
 
-        epic = new Epic("4","");
-        res = taskManager.createEpic(epic);
-        if (res.isPresent())
-            throw res.get();
         subTask = new SubTask("task2","desc", epic.getTaskId());
         subTask.setStartTime(LocalDateTime.of(2024,12,12,2,55));
         subTask.setDuration(Duration.ofMinutes(1));
@@ -257,23 +337,24 @@ public class HttpTaskManagerOfEpicsTest {
 
         var response = getResponse(HttpRequest.newBuilder()
                 .GET()
-                .uri(new URI("http://127.0.0.1:8080/epics/prioritized/-1")));
+                .uri(new URI("http://127.0.0.1:8080/subtasks/prioritized/-1")));
 
-        Assertions.assertEquals(200, response.statusCode());
+        Assertions.assertEquals(200, response.statusCode(), response.body());
 
-        List<Epic> prioritizedList = gson.fromJson(response.body(), new TypeToken<List<Epic>>() {}.getType());
-        var names = new String[] {"2","1","3","4"};
+        List<SubTask> prioritizedList = gson.fromJson(response.body(), new TypeToken<List<SubTask>>() {}.getType());
+        var hours = new Integer[] {6,5,3,2};
         var index = 0;
 
         for (var item : prioritizedList) {
-            Assertions.assertEquals(names[index], item.getName());
+            var expectedVal = hours[index];
+            Assertions.assertTrue(item.getStartTime().filter(i -> i.getHour() == expectedVal).isPresent());
             index++;
         }
     }
 
     @Test
     void badRequest() throws Exception {
-        var epic = new Epic("1","");
+        var epic = new Epic("","");
         var res = taskManager.createEpic(epic);
         if (res.isPresent())
             throw res.get();
@@ -285,10 +366,6 @@ public class HttpTaskManagerOfEpicsTest {
         if (res.isPresent())
             throw res.get();
 
-        epic = new Epic("2","");
-        res = taskManager.createEpic(epic);
-        if (res.isPresent())
-            throw res.get();
         subTask = new SubTask("task2","desc", epic.getTaskId());
         subTask.setStartTime(LocalDateTime.of(2024,12,12,6,55));
         subTask.setDuration(Duration.ofMinutes(1));
@@ -296,10 +373,11 @@ public class HttpTaskManagerOfEpicsTest {
         if (res.isPresent())
             throw res.get();
 
-        epic = new Epic("3","");
+        epic = new Epic("","");
         res = taskManager.createEpic(epic);
         if (res.isPresent())
             throw res.get();
+
         subTask = new SubTask("task1","desc", epic.getTaskId());
         subTask.setStartTime(LocalDateTime.of(2024,12,12,3,5));
         subTask.setDuration(Duration.ofMinutes(1));
@@ -307,10 +385,6 @@ public class HttpTaskManagerOfEpicsTest {
         if (res.isPresent())
             throw res.get();
 
-        epic = new Epic("4","");
-        res = taskManager.createEpic(epic);
-        if (res.isPresent())
-            throw res.get();
         subTask = new SubTask("task2","desc", epic.getTaskId());
         subTask.setStartTime(LocalDateTime.of(2024,12,12,2,55));
         subTask.setDuration(Duration.ofMinutes(1));
@@ -320,7 +394,7 @@ public class HttpTaskManagerOfEpicsTest {
 
         var response = getResponse(HttpRequest.newBuilder()
                 .GET()
-                .uri(new URI("http://127.0.0.1:8080/epics/bad")));
+                .uri(new URI("http://127.0.0.1:8080/subtasks/bad")));
 
         Assertions.assertEquals(500, response.statusCode(), response.body());
     }
