+++ test/taskManagers/TaskManagerFactoryTest.java
@@ -1,4 +1,4 @@
-package tests.taskManagers;
+package taskManagers;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
