+++ test/tasks/SortedTaskMapTest.java
@@ -1,4 +1,4 @@
-package tests.tasks;
+package tasks;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
