+++ test/tools/HistoryManagerTest.java
@@ -1,4 +1,4 @@
-package tests.tools;
+package tools;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
